query-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorlog==6.10.1
datasets==2.14.7
dill==0.3.7
distro==1.9.0
dspy-ai==2.4.9
fastapi==0.115.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2023.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jiter==0.12.0
joblib==1.3.2
kuzu==0.6.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
multiprocess==0.70.15
nats-py==2.12.0
numpy==2.3.4
openai==1.109.1
optuna==4.6.0
packaging==25.0
pandas==2.3.3
propcache==0.4.1
pyarrow==22.0.0
pyarrow-hotfix==0.7
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.38.6
structlog==25.5.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.32.0
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
